# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sabaik Almasa API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: getServices
      tags: [content]
      summary: List services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Containers ────────────────────────────────────────────────────────────
...
  /values:
    get:
      operationId: getValues
      tags: [content]
      summary: List company values
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyValue"
    post:
      operationId: createValue
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyValueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyValue"
  /values/{id}:
    patch:
      operationId: updateValue
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyValueInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyValue"
    delete:
      operationId: deleteValue
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Testimonials ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: admin-token
  schemas:
...
    Service:
      type: object
      required: [id, title, description, icon, order, isActive]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        icon: { type: string }
        imageUrl: { type: ["string", "null"] }
        images: { type: string, description: "JSON array of image URLs" }
        order: { type: integer }
        isActive: { type: boolean }
        seoEnabled: { type: boolean }
        seoTitle: { type: string }
        seoDescription: { type: string }
        seoKeywords: { type: string }
        seoSlug: { type: string }
    ServiceInput:
      type: object
      required: [title, description, icon, order]
      properties:
        title: { type: string }
        description: { type: string }
        icon: { type: string }
        imageUrl: { type: ["string", "null"] }
        images: { type: string, description: "JSON array of image URLs" }
        order: { type: integer }
        isActive: { type: boolean }
        seoEnabled: { type: boolean }
        seoTitle: { type: string }
        seoDescription: { type: string }
        seoKeywords: { type: string }
        seoSlug: { type: string }
...
    CompanyValue:
      type: object
      required: [id, title, description, icon, order]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        icon: { type: string }
        order: { type: integer }
    CompanyValueInput:
      type: object
      required: [title, description, icon, order]
      properties:
        title: { type: string }
        description: { type: string }
        icon: { type: string }
        order: { type: integer }
...